Fix library version and SONAME
With the switch from autotools to CMake the library versioning was not adapted
correctly. (a11a5fb6c466f4a28a35462ac3d3a5725b6e81a0)

For autotools @-version-info N:N:N@ was used to set the library version, for
CMake the library version is incorrectly set as the SOVERSION instead. The
SOVERSION is an incrementing integer and is configured separately from the
library version.